// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  BOOKING_CONFIRMED = 'BOOKING_CONFIRMED',
  BOOKING_CANCELLED = 'BOOKING_CANCELLED',
  BOOKING_REMINDER = 'BOOKING_REMINDER',
  FRIEND_REQUEST_RECEIVED = 'FRIEND_REQUEST_RECEIVED',
  FRIEND_REQUEST_ACCEPTED = 'FRIEND_REQUEST_ACCEPTED',
  NEW_FOLLOWER = 'NEW_FOLLOWER',
  NEW_LIKE = 'NEW_LIKE',
  NEW_COMMENT = 'NEW_COMMENT',
  TREK_UPDATE = 'TREK_UPDATE',
  ORGANIZER_APPROVED = 'ORGANIZER_APPROVED',
  ORGANIZER_REJECTED = 'ORGANIZER_REJECTED',
  STORY_EXPIRING = 'STORY_EXPIRING',
  CHECK_IN_REMINDER = 'CHECK_IN_REMINDER',
  SAFETY_ALERT = 'SAFETY_ALERT',
  EMERGENCY_ESCALATION = 'EMERGENCY_ESCALATION',
  GROUP_INVITE = 'GROUP_INVITE',
  GROUP_UPDATE = 'GROUP_UPDATE',
  WISHLIST_PRICE_DROP = 'WISHLIST_PRICE_DROP',
  BROADCAST = 'BROADCAST',
}